Move create_provider_tokens_object to auth.py and delete conversation_service.py
The function was only used within enterprise/server/routes/auth.py, so it's
been moved there and the now-empty conversation_service.py has been deleted.
